💄 (device-status): Prevent device status blink
