+++ apps/mobile/src/components/DeviceStateView.tsx
@@ -1,17 +1,17 @@
 import React from "react";
 import { Text } from "@ledgerhq/native-ui";
 import { useDeviceSessionState } from "_hooks/useDeviceSessionState";
-import useDebounce from "_hooks/useDebounce";
+import useThrottle from "_hooks/useThrottle";
 
 export const DeviceStateView: React.FC<{ sessionId: string }> = ({
   sessionId,
 }) => {
   const state = useDeviceSessionState(sessionId);
-  const debouncedState = useDebounce(state);
+  const throttledState = useThrottle(state);
 
   return (
     <Text fontWeight="semiBold">
-      Device state: {debouncedState?.deviceStatus}
+      Device state: {throttledState?.deviceStatus}
     </Text>
   );
 };
