--- apps/mobile/src/hooks/useDebounce.ts
@@ -1,19 +1,0 @@
-import { useState, useEffect } from "react";
-
-function useDebounce<T>(value: T, delay: number = 500) {
-  const [debouncedValue, setDebouncedValue] = useState(value);
-
-  useEffect(() => {
-    const handler = setTimeout(() => {
-      setDebouncedValue(value);
-    }, delay);
-
-    return () => {
-      clearTimeout(handler);
-    };
-  }, [value, delay]);
-
-  return debouncedValue;
-}
-
-export default useDebounce;
