+++ apps/mobile/src/hooks/useThrottle.ts
@@ -1,0 +1,19 @@
+import { useState, useEffect } from "react";
+
+function useThrottle<T>(value: T, delay: number = 500) {
+  const [throttledValue, setThrottledValue] = useState(value);
+
+  useEffect(() => {
+    const handler = setTimeout(() => {
+      setThrottledValue(value);
+    }, delay);
+
+    return () => {
+      clearTimeout(handler);
+    };
+  }, [value, delay]);
+
+  return throttledValue;
+}
+
+export default useThrottle;
