+++ apps/sample/src/hooks/useThrottle.ts
@@ -1,0 +1,17 @@
+import { useEffect, useState } from "react";
+
+export function useThrottle<T>(value: T, delay: number = 500) {
+  const [throttledValue, setThrottledValue] = useState(value);
+
+  useEffect(() => {
+    const handler = setTimeout(() => {
+      setThrottledValue(value);
+    }, delay);
+
+    return () => {
+      clearTimeout(handler);
+    };
+  }, [value, delay]);
+
+  return throttledValue;
+}
